+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - CockroachDB: Treat crdb_internal and pg_extension as read-only system schemas, e.g. do not export them
 - MS SQL: Export: Create the schema by a new option
 - MS SQL: Fix exporting more schemas at once
+- MS SQL: Fix the length of a varchar(max) column
 - Oracle: The driver is no longer beta
 - Oracle: Use schemas as databases instead of tablespaces (bug SF-204, SF-212, SF-238)
 - Oracle: Create and drop databases (bug SF-246)